+++ NEWS.md
@@ -1,4 +1,4 @@
-# branch pol-escapa #
+# branch Pol-Escapa #
 
 - New function `PolEscapa()` tests which characters contribute to taxon
   instability, per Pol & Escapa (2009) <doi:10.1111/j.1096-0031.2009.00258.x>.